requirements.txt
altair==5.1.2
asgiref==3.7.2
aspose-cells==23.10.0
asttokens==2.4.0
attrs==23.1.0
bcrypt==4.0.1
beautifulsoup4==4.12.2
behave==1.2.6
bitsandbytes==0.41.1
blinker==1.6.3
Brotli==1.1.0
cachetools==5.3.2
certifi==2023.7.22
cffi==1.16.0
chardet==5.2.0
charset-normalizer==3.3.0
chromedriver-autoinstaller==0.6.2
click==8.1.7
colorama==0.4.6
contourpy==1.1.1
cryptography==41.0.4
cssselect==1.2.0
cycler==0.12.1
devtools==0.12.2
Django==4.2.6
EasyProcess==1.1
entrypoint2==1.1
entrypoints==0.4
et-xmlfile==1.1.0
exceptiongroup==1.1.3
execnet==2.0.2
executing==2.0.0
extra-streamlit-components==0.1.60
Faker==19.12.0
fasteners==0.19
favicon==0.7.0
filelock==3.12.4
Flask==3.0.0
fonttools==4.43.1
gitdb==4.0.11
GitPython==3.1.40
h11==0.14.0
h2==4.1.0
hpack==4.0.0
htbuilder==0.6.2
hyperframe==6.0.1
idna==3.4
image==1.5.33
importlib-metadata==6.8.0
iniconfig==2.0.0
itsdangerous==2.1.2
Jinja2==3.1.2
JPype1==1.4.1
jsonschema==4.19.2
jsonschema-specifications==2023.7.1
kaitaistruct==0.10
kiwisolver==1.4.5
lxml==4.9.3
Markdown==3.5
markdown-it-py==3.0.0
markdownlit==0.0.7
MarkupSafe==2.1.3
matplotlib==3.8.0
mdurl==0.1.2
more-itertools==10.1.0
MouseInfo==0.1.3
mss==9.0.1
numpy==1.26.1
opencv-python==4.8.1.78
openpyxl==3.1.2
outcome==1.3.0
packaging==23.2
pandas==2.1.2
parameterized==0.9.0
parse==1.19.1
parse-type==0.6.2
pdbp==1.5.0
Pillow==10.1.0
platformdirs==3.11.0
plotly==5.18.0
pluggy==1.3.0
protobuf==4.24.4
py==1.11.0
pyarrow==13.0.0
pyasn1==0.5.0
PyAutoGUI==0.9.54
pycparser==2.21
pydeck==0.8.1b0
pydivert==2.1.0
PyGetWindow==0.0.9
Pygments==2.16.1
PyJWT==2.8.0
pymdown-extensions==10.3.1
PyMsgBox==1.0.9
pynose==1.4.8
pyOpenSSL==23.2.0
pyotp==2.9.0
pyparsing==3.1.1
pyperclip==1.8.2
pyreadline3==3.4.1
PyRect==0.2.0
pyscreenshot==3.1
PyScreeze==0.1.29
PySocks==1.7.1
pyTelegramBotAPI==4.14.0
pytest==7.4.2
pytest-html==2.0.1
pytest-metadata==3.0.0
pytest-ordering==0.6
pytest-rerunfailures==12.0
pytest-xdist==3.3.1
python-dateutil==2.8.2
python-dotenv==1.0.0
pytweening==1.0.7
pytz==2023.3.post1
pywhatkit==5.4
PyYAML==6.0.1
referencing==0.30.2
requests==2.31.0
rich==13.6.0
rpds-py==0.10.6
sbvirtualdisplay==1.2.0
six==1.16.0
smmap==5.0.1
sniffio==1.3.0
sortedcontainers==2.4.0
soupsieve==2.5
sqlparse==0.4.4
st-annotated-text==4.0.1
streamlit==1.28.0
streamlit-authenticator==0.2.3
streamlit-autorefresh==1.0.1
streamlit-camera-input-live==0.2.0
streamlit-card==0.0.61
streamlit-embedcode==0.1.2
streamlit-extras==0.3.4
streamlit-faker==0.0.3
streamlit-image-coordinates==0.1.6
streamlit-keyup==0.2.0
streamlit-option-menu==0.3.6
streamlit-toggle-switch==1.0.2
streamlit-vertical-slider==1.0.2
tabcompleter==1.3.0
telebot==0.0.5
tenacity==8.2.3
toml==0.10.2
toolz==0.12.0
tornado==6.3.3
tradingview-ta==3.3.0
trio==0.22.2
trio-websocket==0.11.1
typing_extensions==4.8.0
tzdata==2023.3
tzlocal==5.2
undetected-chromedriver==3.1.5.post4
urllib3==1.26.18
urllib3-secure-extra==0.1.0
validators==0.22.0
watchdog==3.0.0
webdriver-manager==4.0.1
websockets==11.0.3
Werkzeug==3.0.0
wget==3.2
wikipedia==1.4.0
wsproto==1.2.0
zipp==3.17.0
zstandard==0.21.0
